Uploaded updated drawbox.py and added drawbox_random.py.
drawbox.py is updated to work on the new dataset structure. Also added drawbox_random.py which randomly opens any file in given datasets and shows the image with the bounding boxes. If any new dataset is in yolo format (structured like e.g. the deepdrive dataset) both codes should be easyly maintainable with only adding the new dataset to the dataset list in the file. It' great to find out if the conversion of labels into the yolo format was successful.